// src/app/models/signup.model.ts
export class Signup {

    key?: string | null;
    fullName?: string;
    UTDEmail?: string;
    joinDate?: string;
    phoneNumber?: string;
    major?: string;
    ToLearnProgLanuages?: string;
    Hackathons?: string;
    Networking?: string;
    ToBecomeAnOfficer?: string;
    AreYouAnOfficer?: string;
    Team?: string;
    txnId?: string;
}